+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Start configed when trusting the certificate once
 - Nullpointerexception when user is not in group opsi admin
 - Correctly fix default values as specific values for product property states on clients and depots
+- Correctly deactivate filter in client selection
 
 ### Changed
 - Move reload and save buttons to top toolbar